move prep code into common
