+++ src/plopp/functions/slicer.py
@@ -1,9 +1,9 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright (c) 2022 Scipp contributors (https://github.com/scipp)
 
+from .common import is_interactive_backend, preprocess
 from .figure import figure
 from ..model import input_node, widget_node
-from ..prep import preprocess
 
 from scipp import Variable, Dataset
 from scipp.typing import VariableLike
@@ -43,7 +43,7 @@ def slicer(obj: Union[VariableLike, ndarray],
     :
         A :class:`Box` which will contain a :class:`Figure` and slider widgets.
     """
-    if not _is_interactive_backend():
+    if not is_interactive_backend():
         raise RuntimeError("The slicer can only be used with the interactive widget "
                            "backend. Use `%matplotlib widget` at the start of your "
                            "notebook.")
